Updated logic for Container and AD.
Now users and groups are only stored in containers (except them referencing
each other when necessary). Containers are only stored in directories. Users
and groups are not directly referenced in directories. That was just too
complicated. It makes sense for users and groups to reference each other
however, because users can have groups in different containers (and vice-versa).