Added RID handling and User primary_group attribute and handling.
When loading users and groups from Active Directory, a group's member attribute
contains a list of users who are members. However, group membership can also
be defined by the user's primaryGroupID attribute - which Microsoft has
"conveniently" stored as a RID value. Of course, there is no way to easily
find the RID of any object in Active Directory because the RID is at the
end of an object's SID (even though there seems to be an empty rid attribute
that isn't used). What is up with all this? Well, the code here was created
to deal with this so that loading users and groups from a directory would
work correctly. This also means that users need the concept of "primary_group",
which corresponds to microsoft's attempt to implement the POSIX requirement
that a user has to have a primary group (Windows doesn't care otherwise, see
the GUI tool where you can modify this - it says, in effect, "Don't bother!").
The tests were updated as well. Keep in mind that RIDs are in a flat namespace.