Added primary_group and unix_main_group logic changes.
The Container class does not remove a group if it is a any user's
primary_group or unix_main_group. The initialize() method of the User and
UNIXUser classes now take a primary_group argument. The groups array of the
User/UNIXUser class contains the unix_main_group, but not the primary_group
as users are implicitly members of their primary_group. The member_of?()
method was modified to reflect this. The full_name, password, and gecos
attributes of users now have default values. The primary_group cannot be
added to the groups array of a user. Also, main_group was changed to
unix_main_group to be less confusing. Setting the unix_main_group attribute
of a UNIXUser was updated to handle the new logic where that group is added
to the user's groups array. The Group class will not add a user if that user
has the group as its primary_group. Note: in all cases, the groups/users array
reflects the AD related attributes, so this all makes sense. A users array
representation was added to the AD class to make it easier to detect if
a group is some user's primary_group or unix_main_group in the Container
class for its check (mentioned first here). Updated tests to match new logic.