Added notes and user disabled logic.
I added some notes about what I've found with respect to the group scope and
type in AD as well as whether an account is disabled or enabled. All I really
care about is the disabled value, which seems consistent. There are probably
other values for "enabled", but the dsmod command has a specific flag for
disabled so I can use this simplified logic.

I allow an account to be disabled or enabled, but not changing the group
type or scope because there's a lot more logic there (I don't know if that'd
really be used - I might add it later). For now, create the groups
appropriately upfront :-)