Renamed the module to RADUM.
The module has been renamed to RADUM (Ruby Active Directory User Management).
This is not meant to be a general Active Directory tool. It is meant to
faciliate user management, specifically Windows users that might also,
and probably will, have UNIX attributes. The same goes for groups. I want
this module to do as many things with those objects as possible automatically,
instead of providing a lower level interface to Active Directory and requiring
the user to do all the logic. I guess it can be viewed as more and less
powerful at the same time, depending on what you want.

There are at least two "active-directory" type modules/Gems already. This is
not meant to compete wtih those packages really, and this avoids a name
collision.