Setting the nis_domain to some useful default value.
The nis_domain attribute should be set to something. This is not strictly
needed if one is just storing UNIX attributes in AD, but it is necessary
in order to edit those attributes using the AD Users and Groups tool.
Windows expects to be set to something. Even if you are not using NIS
directly, setting this attribute just makes things easier.