Updated notes and RDoc.
Updated Notes.txt with new information and details about class attribute
mappings to LDAP attributes. Made some RDoc fixes. Finished RDoc for the
User class and started on the UNIXUser class. Add additional class attributes
to User and UNIXUser to set everything needed (so far). Started the convention
of using nil for attributes that will not be set in AD when updating or
creating entries. Updated the AD.load method to handle new attributes and
fixed the checks for NoMethodError to make sure every attribute possible is
found.