Added the modified attribute.
A modified attribute was added to User, UNIXUser, Group and UNIXGroup classes.
The modified attribute is generally read-only. When new objects are created
by hand, the modified attribute is set to true. There is a loaded method
that will set it to false if called last in AD.load (it has to be last). The
loaded method can be called only once for effect. It sets the loaded attribute
and resets the modified attribute to false. Various changes will set modified
to true when needed - this is why loaded must be called last. That's all
right since it doesn't really affect clients (just me).

This required rewriting the accessors from their shorcut forms in order
to set the modified attribute as well. It is possible to do more checks, but
for now I am setting the modified attribute even if the user sets the same
value. The point is to keep from trying to sychronize objects that have not
changed. I've not done that part yet, but this is a first stab at keeping
unnecessary syncs from happening. We'll see :-)