Finally able to create a user using LDAP alone.
This is apparently difficult. It took forever for me to figure out why I
was getting error code 53 (unwilling to perform) when trying to do this.
First the user has to be created with no password and disabled, then you
can try setting a password. I am checking this in because I want to do
some other fixes too. It is nowhere near production. I am just so *happy*
I got this far...