Added ASCII to UTF-16LE method for Windows unicodePwd attribute.
Windows expects the unicodePwd attribute to be an explicitly quoted UTF-16LE
string. This is easy to do since the string should simply be each character
followed by a NULL. It took forever to figure this out (not really, but
it feels like it :-). There is still a ton to do, but this is good progress.
Note that the AD object requires SSL or the Windows account cannot be enabled
and have its password set.