Updated user creation.
User creation now handles the Windows password better. I also cleaned it up
and added a method to check the LDAP results to dry things up a little. This
needs a lot of work however. UNIX attributes are handled now as well as setting
the user's primary Windows group correctly.