Added Container creation during synchronization.
Container objects might need to be created as well, but I am treating them as
if they are in a flat namespace. Containers don't directly hold other Container
objects. It could be done that way, but I would rather have the ability to
work with one subcontainer as desired. This all works fine because of how
the Container class name attribute includes the type (cn for contianer or
ou for organizational unit - the only two types supported). When creating
containers, it's possible to do some magic with the name using split(), etc.
to create parent entities if necessary implicitly. Did I mention I wanted
things to be magictastical in RADUM?