First step toward syncing user modifications.
Changed the code to facilitate synchronizing user updates. This included
have some things updated after user creation as well. There is a lot of
work to do. I tried to dry things up a little, but this is fairly complicated
because I am trying to hide certain things and enforce various rules instead
of making the user do that.