Added random password generation and updated user synchronizing.
I added code to handle generating a random password for a new user if a
password was not specified using a format I think will work most of the
time. If not, the user can specify a password they know will work. I
updated the logic so that password changes are not forced on the next
login by default. I fixed some code formatting issues (extra space found
its way in there). After the password is set in Active Directory, a User
or UNIXUser has its password attribute set to nil again. This is because
the password attribute is only to be set to some value to cause it to
change. It does not reflect the user's actual Windows password otherwise
because that cannot be read from LDAP directly. Added more code to the
update_user method to handle a password change and forcing a user to change
their password.