Added to user updating, group updating, and setting the primary group.
Added logic that makes sure changing the User or UNIXUser object's
primary_group also adds standard membership in the old primary_group. This
reflects what happens when the primary_group is changed in Windows. I also
added a test to make sure that happens. I fixed a small bug in marking a
user as loaded in the create_user() method. I added handling the disabled?
attribute in update_user. I also started working on the update_group() method,
which is much simpler than the update_user() method. In both cases, I added
a new check to make sure that only attributes in the attr hash that apply
are checked. The check is needed because the *_ldap_entry_attr() methods
fill in every possible value with something, which works for UNIXUser and
UNIXGroup objects - but it is too much for User and Group objects. The
*_ldap_entry_attr() methods don't have a concept for the different types,
so anything not found is set to nil. This makes logical sense to me, but
it has to be dealt with wherever the attr has returned is used.