Added user and group removal.
Added the first sync() steps that remove users, then remove groups. Users
need to be removed first in case any of them have a removed group as their
primary Windows group. There still needs to be a check for any main UNIX
group members as well. The search in remove_group() needs to check all
users in Active Directory since some might not be known about in the RADUM
system at any given point. This is inefficient, but removing groups is not
a usual action in my experience. Even so, it should not be too bad.