Expanded Container removal.
Added code that handles Container removal properly. All users are also removed,
all groups are removed (if possible from what is known), and all Containers
are checked for a Container inside the target Container. If none of these fail,
the Container itself is removed. If one condition cannot be met, everything that
can be removed is removed. I renamed the private AD methods for removal to use
the word "delete" instead so that it is clearer and so there is no conflict
with the public remove_container method. I fixed a bug with the Container
remove_group() method.