Updated User/UNIXUser new() method arguments.
Updated the User and UNIXUser new methods to take a hash of arguments. This
should be easier to deal with in the code. Fixed a bug where the initialize()
method for the UNIXUser class would always attempt to make the user a member
of their unix_main_group even if it was equal to their primary_group. That
should not happen.