Updated AD#load and AD#sync logic.
It is now possible to load multiple times. Anything previously loaded will
be ignored. Calling AD#sync is smoother and safe (which is probably was before)
to call more than once. There is logic to destroy things that were removed
from Active Directory properly. The set_loaded() methods were updated so they
can be called after an initial load and do the same thing. The set_loaded
methods are not meant to be called by the consumer, but they need to be there
for the AD object.