Added the "Home folder" part of the Profile tab.
Added a way to set either the "Local path" or "Connect ... To" parts of
the Home folder section of the Profile tab in the Active Directory Users
and Computers tool. This gets a little confusing because these options
are mutually exclusive, but there are two instance variables (and two
LDAP attributes) that represent them. I tried to explain this well, but
I will need to add a serious example. This should work right - provided
others know what I'm talking about based on the documentation.