Fixed minor bugs with AD#load and AD#update_group and added new methods.
I fixed a bug in AD#load where it would not handle managing adding users
to new groups if they do not already exist in Active Directory. It is
possible to create a group before calling AD#load, so we can't depend on
the member attribute actually existing at that point.

In AD#update_group the member and msSFU30PosixMember attributes were not
being updated in the case where they did not already have a value.

I added methods to convert a Group to a UNIXGroup and a UNIXGroup to a Group.
Even if these changes are dropped, the other two fixes are important.