Added middle initials handing for user and fixed dn bug.
Added the middle initials attribute as that is usually what is set instead
of middle_name. I fixed a bug with the distinguished_name attribute. It is
not safe to assume the defaults I had for users who are created with the
GUI tools. More testing has to be done here as well anyway.