Added primary_group tests and fixed primary_group order bug.
Added testing for primary_group changes. In testing I discovered a slight
bug with the order of operations. Since we update the RADUM enviornment to
reflect the intention and groups are updated before users, the code would
try to add the user as a member of their old primary Windows group before
the primary group was changed for the user himself. The code actually worked,
but an LDAP error was returned. Instead, the new code checks if the user
in LDAP has had their primaryGroupID changed yet. If not, and it matches
the group being updated, we don't add them as a member. They will be added
as a member implicitly when AD#update_user() is called.