Updated "removed" logic and updated live test.
I changed the system so that one cannot add objects that were removed or
destroyed back into the system. This also changes the logic in the initialize()
methods. I also was adding the unix_main_group to the Container for some reason
even though it has to already be there anyway.