Modified all removed logic and fixed UNIXGroup bug.
Modified the previous removed attribute logic. It used to be possible to
add removed objects back into the RADUM environment, but there was a problem
with this when it came to the loaded status because adding the object back
would try to update it instead of create it if AD#sync had been called between
the removal and addition back. There is really no use for this feature, so
it was itself removed. If an object is removed, it can't be added back into
the system. I also added many more tests for trying to pass in removed
(and by extension destroyed) objects where I was not checking before. This
caused a lot of other changes in tests and main body code too. In making
the removed changes, I accidently removed setting the unix_password for
UNIXGroup objects, which caused them not to get created properly. That
was fixed.

The live test was updated with the first step of UNIXUser testing.