UNIXUser/UNIXGroup live testing updates and fixes.
- Added an ldap attribute to the AD object for easy external use if needed.

- Fixed a bug with load_next_uid and load_next_gid when UIDs/GIDs were in AD
and Active Directory so appeared twice in the search array built that ensures
UIDs/GIDs not yet syncrhonized are accounted for. The algorithm assumed that
there were only unique values in the search array generated. Now that is the
case.

- Misc. logger message fixes.

- Fixed Active Directory LDAP attribute vs. object attribute value comparisons.
Most values returned from an LDAP query are String values, but they are really
used as integers. In the code they are stored as integers, so the comparison
was not always accurate. This did not cause a problem, except for updates
when not needed (only for the UNIXUser object that had the shadow file
attributes always being updated). Now the comparision is done as String
values. This even works when comparing the primary Windows group since it
has a working to_s() method. There are some other cases where binary values
are returned from Active Directory LDAP queries, but let's pretend I didn't
mention that (those would not be compared by me any time soon).

- Modified the previous primary Windows group delayed membership handling
in AD#update_group() to only check when the object "item" is a User or
UNIXUser object (it could be a Group or UNIXGroup object in that code). This
is a cleaner version that is easier to understand... for me anyway.

- Added code to AD#update_user() to handle UNIX main group changes when
necessary because implicit UNIX group membership rules are a little different
than Windows group memberships. This is an easy way to deal with that problem,
at least with the code the way it is and the rules I put in place.

- Fixed a precedence bug in User#member_of?().

- Updated the live testing script with extensive UNIXUser/UNIXGroup testing,
which is the reason for all of these updates. And here I thought things
were mostly fine before I started testing UNIX stuff!