Fixed scope of primary Windows group check in conversions.
It is safe to convert a group if it is the primary Windows group of a user
RADUM doesn't know about, but not if it is one that RADUM does know about -
because I have to have that logic for other reasons. Only the UNIX main group
full Active Directory check is necessary beyond what Container#destroy_group
would do in this case.