Added group live testing and fixed group handling bug.
Added code to make sure the user does not lose their UNIX membership
in a group when it is made the user's primary Windows group. Before
the Windows membership and UNIX membership was removed. If there
was a previous UNIX group membership, it should not be removed in that
step because the user should still be a UNIX member along with their
implicit Windows membership.

Added code to add a group to a group. This was not well tested and was
not working from the Active Directory perspective.

The code to check if a group is a member of another group was reversed
originally, thus it was fixed.

Modified group remove_user() code to ignore the case where the UNIX main
group is trying to be removed if it is also the primary Windows group.
In that case, we want the group to be removed from the user's groups array
even though it is the UNIX main group as the code handles that internally.

Changed the order of operations in primary_group= for users so that the
primary_group is set before trying to remove_group() the group so that
the UNIXUser will allow removal even if the group is the current UNIX
main group (because it can be determined that it is also the primary
Windows group - thus the special case is caught). This is due to internal
logic handling implicit memberships from the Windows and UNIX perspective.

Added comments to the remove_group() for UNIXUser objects that says it
will not raise the "group is UNIX main group" RuntimeError if the group
is also the primary Windows group to handle the special case where the
group is both.

Updated live testing to test group aspects. This should be about all there
is.