Improved white space handling and fixed delete_container().
Improved the white space handling. Now Container objects can have spaces
in their names, but extraneous white space is removed. White space is stripped
from the username and name arguments for User/UNIXUser and Group/UNIXGroup
objects. The comments were updated to reflect these improvements. The
delete_container() method in the AD object only deletes a Container if it
actually is in Active Directory instead of just resulting in an LDAP error.
That is better behavior.