Fixed a bug with UNIXGroup updates and with "destroy" logic.
The AD#unix_group_to_group method needed to delete the gidNumber LDAP UNIX
attribute. The AD#update_group method needed to set the GID for newly
converted to UNIXGroup objects. In testing I discovered that the "destroy"
methods were not working how I intended. They were not really removing
references to users and groups - they acted a lot like the "remove" methods
(and in fact, that is how the code used them). However, this would leave old
references in the User/UNIXUser and Group/UNIXGroup object's users/groups
arrays. For example, if you converted a User to a UNIXUser, a group might
still have the old User around because it was not really "destroyed". The
"destroy" logic was updated accordingly, which cased the "remove" logic
to need updating.

In AD#sync I also noticed that I could not destroy users and containers in
the AD#delete_user and AD#delete_container methods. The AD#update_group()
method still needs references to all Container objects to find out who is
no longer in a group via being destroyed. Removing a user from Active Directory
impliciltly takes care of Windows group memberships, but AD#update_group
still needs to know about the objects that were deleted in order to update
the UNIX group membership attributes on what's left (since that's not updated
implicitly).