Updated AD#destroy_container logic, comments, tests, etc.
- Expaned AD test suite to test new AD#remove_container and
AD#destroy_container logic. This tests all the other remove/destroy logic
as well through side effects.

- Updated comments for delayed destroy logic after the previous commit
related to that.

- Updated AD#destroy_container to only destroy a Container if possible.

- Removed duplicate user destruction from UNIX main group (should already
do that anyway).

- Removed extra push of user to removed_users array for the UNIX main group
if it was also the user's primary Windows group. The user should not be
pushed to the removed_users array in that case because they do not have
a regular Windows membership there to remove (was a member just because
of primaryGroupID attribute for the user and their UNIX membership was
only through their gidNumber attribute - nothing to do in that case when
the user is removed).