Fixed a bug with UNIX main group changes and added test case.
Fixed a bug with UNIX main group changes. If the UNIX main group is changed
and the UNIXUser is explicitly removed from the old UNIX main group, they
would end up with a membership from the UNIX perspective only. Now that
won't happen. I added a test case that will find this bug.