Fixed Ruby 1.9 frozen string problem and added explicit Ruby requirement.
Ruby 1.8.5 or higher is required, so that's been specified. It might work
with versions less than Ruby 1.8.5, but I have not tested it. RADUM will
not work with Ruby 1.9 yet because of ruby-net-ldap, but I did fix a frozen
String problem when using the ENV hash values for later Ruby 1.9 work.