Added notes about the Certificate Authority Web Enrollment role service.
Mac OS X Snow Leopard requires the Certificate Authority Web Enrollment
role service to be installed on the Active Directory server. This took
me a while to figure out. If that role service is not installed, RADUM
will be _extremely_ slow (totally unusable). This service is probably
a good idea anyway of course. I added a note about that in the comments.